Day 15, part 2
So I saved the map that I manually built from part one, then used that
for part 2, rather than try and build it automatically. I guess that's
cheating a little.

As for the actual solution, I remembered a similar problem from last
year's AoC.

In fact, I could probably use the same approach to solve part one (same,
but in reverse). But I'd still need to figure how to 'reveal' the map
automatically